# backend/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.CreateSessionRequest:
    properties:
      args:
        items:
          type: string
        type: array
      command:
        type: string
      memberId:
        type: string
      memberName:
        type: string
      terminalType:
        type: string
      workspaceId:
        type: string
    type: object
  handlers.CreateSessionResponse:
    properties:
      pid:
        type: integer
      sessionId:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Multi-Agent Collaboration Platform API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Orchestra API
  version: "1.0"
paths:
...
  /api/terminals:
    post:
      consumes:
      - application/json
      description: Create a new PTY terminal session for an AI assistant or shell
      parameters:
      - description: Session configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateSessionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.CreateSessionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create terminal session
      tags:
      - terminals
  /api/terminals/{sessionId}:
    delete:
      description: Terminate and remove a terminal session
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete terminal session
      tags:
      - terminals
...
  /api/workspaces/{id}/members/{memberId}/terminal-session:
    get:
      description: Get the active terminal session for a workspace member
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      - description: Member ID
        in: path
        name: memberId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CreateSessionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get member's terminal session
      tags:
      - terminals
    post:
      consumes:
      - application/json
      description: Get existing terminal session or create a new one for the member
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      - description: Member ID
        in: path
        name: memberId
        required: true
        type: string
      - description: Optional session configuration
        in: body
        name: request
        schema:
          $ref: '#/definitions/handlers.CreateSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CreateSessionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get or create member's terminal session
      tags:
      - terminals
...
  /api/workspaces/{id}/terminal-sessions:
    get:
      description: Get all active PTY sessions for a workspace
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List workspace terminal sessions
      tags:
      - terminals
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"